requirements-test.txt
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-xdist==3.3.1
coverage==7.3.2
factory-boy==3.3.0
freezegun==1.2.2
